Use Azure SSH key resources
