+++ package.json
@@ -312,8 +312,8 @@
         "webpack-cli": "^3.3.0"
     },
     "dependencies": {
-        "@azure/arm-compute": "^14.0.0",
-        "@azure/arm-network": "^22.0.0",
+        "@azure/arm-compute": "^15.0.0",
+        "@azure/arm-network": "^23.0.0",
         "fs-extra": "^8.1.0",
         "semver": "^5.7.0",
         "vscode-azureextensionui": "^0.37.0",
