+++ src/commands/createVirtualMachine/VirtualMachineCreateStep.ts
@@ -11,7 +11,6 @@ import { ext } from '../../extensionVariables';
 import { localize } from '../../localize';
 import { createComputeClient } from '../../utils/azureClients';
 import { nonNullProp, nonNullValueAndProp } from '../../utils/nonNull';
-import { getSshKey } from "../../utils/sshUtils";
 import { IVirtualMachineWizardContext } from './IVirtualMachineWizardContext';
 import { VirtualMachineOS } from './OSListStep';
 
@@ -40,8 +39,7 @@ export class VirtualMachineCreateStep extends AzureWizardExecuteStep<IVirtualMac
         const linuxConfiguration: ComputeManagementModels.LinuxConfiguration = {
             disablePasswordAuthentication: true, ssh: {
                 publicKeys: [{
-                    // tslint:disable-next-line: strict-boolean-expressions
-                    keyData: await getSshKey(vmName, context.passphrase || ''),
+                    keyData: context.sshPublicKey,
                     // because this is a Linux VM, use '/' as path separator rather than using path.join()
                     path: `/home/${context.adminUsername}/.ssh/authorized_keys`
                 }]
